requirements.txt
autopep8==1.5.2
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
dnspython==1.16.0
idna==2.9
pycodestyle==2.6.0
pymongo==3.10.1
python-dotenv==0.13.0
requests==2.23.0
soupsieve==2.0
urllib3==1.25.9
